feat(core): add wallet security alerts (power down/routes), routes rework, and UI/FAQ improvement
