+++ src/app/assets/stylesheets/_themes.scss
@@ -9,6 +9,8 @@ $themes: (
     backgroundColorOpaque: $color-background-off-white,
     backgroundTransparent: transparent,
     backgroundColorWarning: $color-background-warning,
+    backgroundColorDanger: $alert-color,
+    backgroundColorSecondary: $color-white,
     moduleBackgroundColor: $color-white,
     menuBackgroundColor: $color-background-dark,
     moduleMediumBackgroundColor: $color-white,
@@ -20,6 +22,7 @@ $themes: (
     borderDark: 1px solid $color-text-gray,
     borderAccent: 1px solid $color-blue,
     borderWarning: 1px solid $color-text-warning,
+    borderDanger: 1px solid $alert-color,
     borderTransparent: transparent,
     roundedCorners: 5px,
     roundedCornersTop: 5px 5px 0 0,
@@ -30,6 +33,7 @@ $themes: (
     textColorAccent: $color-text-blue,
     textColorAccentHover: $color-blue-original-dark,
     textColorError: $color-text-red,
+    textColorOnDanger: $color-white,
     contentBorderAccent: $color-transparent,
     buttonBackground: $color-blue-original-dark,
     buttonBackgroundHover: $color-blue-original-light,
@@ -48,6 +52,8 @@ $themes: (
     backgroundColorOpaque: $color-background-off-white,
     backgroundTransparent: transparent,
     backgroundColorWarning: $color-background-warning,
+    backgroundColorDanger: $alert-color,
+    backgroundColorSecondary: $color-white,
     moduleBackgroundColor: $color-white,
     menuBackgroundColor: $color-background-dark,
     moduleMediumBackgroundColor: $color-transparent,
@@ -59,6 +65,7 @@ $themes: (
     borderDark: 1px solid $color-text-gray,
     borderAccent: 1px solid $color-teal,
     borderWarning: 1px solid $color-text-warning,
+    borderDanger: 1px solid $alert-color,
     borderTransparent: transparent,
     roundedCorners: 5px,
     roundedCornersTop: 5px 5px 0 0,
@@ -69,6 +76,7 @@ $themes: (
     textColorAccent: $color-text-teal,
     textColorAccentHover: $color-teal,
     textColorError: $color-text-red,
+    textColorOnDanger: $color-white,
     contentBorderAccent: $color-teal,
     buttonBackground: $color-blue-black,
     buttonBackgroundHover: $color-teal,
@@ -87,6 +95,8 @@ $themes: (
     backgroundColorEmphasis: $color-background-super-dark,
     backgroundColorOpaque: $color-blue-dark,
     backgroundColorWarning: $color-background-warning-dark,
+    backgroundColorDanger: $alert-color,
+    backgroundColorSecondary: $color-blue-dark,
     moduleBackgroundColor: $color-background-dark,
     backgroundTransparent: transparent,
     menuBackgroundColor: $color-blue-dark,
@@ -99,6 +109,7 @@ $themes: (
     borderDark: 1px solid $color-text-gray-light,
     borderAccent: 1px solid $color-teal,
     borderWarning: 1px solid $color-background-warning-dark,
+    borderDanger: 1px solid $alert-color,
     borderTransparent: transparent,
     roundedCorners: 5px,
     roundedCornersTop: 5px 5px 0 0,
@@ -109,6 +120,7 @@ $themes: (
     textColorAccent: $color-teal,
     textColorAccentHover: $color-teal-light,
     textColorError: $color-text-red,
+    textColorOnDanger: $color-white,
     contentBorderAccent: $color-teal,
     buttonBackground: $color-white,
     buttonBackgroundHover: $color-teal,
