+++ src/app/client_config.js
@@ -51,3 +51,7 @@ export const SITE_DESCRIPTION =
 
 // various
 export const SUPPORT_EMAIL = 'support@' + APP_DOMAIN;
+
+// External links
+export const STEEMSCAN_BLOCK_URL = "https://steemscan.com/block";
+export const STEEMSCAN_TRANSACTION_URL = "https://steemscan.com/transaction";
