+++ src/app/components/all.scss
@@ -29,6 +29,8 @@
 @import './elements/OutgoingDelegations';
 @import './elements/VotersModal';
 @import './elements/ProposalCreatorModal';
+@import './elements/RouteSettings';
+@import './elements/WithdrawRoutesModal';
 
 // modules
 @import './modules/Header/styles';
