+++ src/app/components/elements/ConversionsModal.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactModal from 'react-modal';
 import CloseButton from 'app/components/elements/CloseButton';
+import tt from 'counterpart';
 
 ReactModal.defaultStyles.overlay.backgroundColor = 'rgba(0, 0, 0, 0.6)';
 
@@ -9,22 +10,22 @@ const ConversionsModal = ({ isOpen, onClose, combinedConversions }) => {
         <ReactModal
             isOpen={isOpen}
             onRequestClose={onClose}
-            className="VotersModal__content"
+            className="ContainerModal__content"
             ariaHideApp={false}
         >
             <CloseButton onClick={onClose} />
             <div className="conversions-modal__container">
-                <h3>Conversions</h3>
+                <h3>{tt('converttosteem_jsx.current_conversions')}</h3>
                 {combinedConversions.length === 0 ? (
-                    <p>No conversion data available.</p>
+                    <p>{tt('converttosteem_jsx.no_conversion_data')}</p>
                 ) : (
                     <table className="conversions-table">
                         <thead>
                             <tr>
-                                <th>ID</th>
-                                <th>Request ID</th>
-                                <th>Amount</th>
-                                <th>Date</th>
+                                <th>{tt('converttosteem_jsx.id')}</th>
+                                <th>{tt('converttosteem_jsx.request_id')}</th>
+                                <th>{tt('g.amount')}</th>
+                                <th>{tt('g.date')}</th>
                             </tr>
                         </thead>
                         <tbody>
