+++ src/app/components/elements/OutgoingDelegations.scss
@@ -2,11 +2,20 @@
     width: 100%;
     font-family: "Source Sans Pro", "Helvetica Neue", Helvetica, Arial, sans-serif;
     font-size: 1rem;
-    border: 1px solid #7e7d7d;
     outline: none;
     padding: 0.25rem 0.5rem 0.5rem 0;
     border-radius: 5px;
     padding: 1.5rem .5rem;
+    border-color: transparent;
+    background: transparent;
+    box-shadow: 0 0 5px rgba(109, 207, 246, 0.5);
+    @include themify($themes) {
+     color: themed('textColorPrimary');
+   }
+  input::placeholder {
+    font-weight: lighter;
+    color: #7e7d7d;
+  }
 }
 
 .input-box{
@@ -16,30 +25,5 @@
 
 .input-box input:focus{
     border: 1px solid #06D6A9;
-}
-
-.input-box label {
-    transform: translateX(10px) translateY(-3rem);
-    left: 0;
-    padding: 1rem;
-    padding-top: 0.5rem;
-    pointer-events: none;
-    font-size: 1rem;
-    color: #7e7d7d;
-    transition: 0.5s;
-    margin-right: 15px;
-    max-width: fit-content;
-}
-
-.input-box input:focus ~ label,
-.input-box .focus ~ label{
-    transform: translateX(10px) translateY(-3.95rem);
-    font-size: .75rem;
-    padding: .45rem;
-    padding-top: .1rem;
-    padding-bottom: .1rem;
-    background-color: #06D6A9;
-    font-weight: bold;
-    color: #FFFFFF;
-    border-radius: 5px;
+    background: transparent;
 }
