+++ src/app/components/elements/UserKeys.jsx
@@ -456,13 +456,23 @@ class UserKeys extends Component {
                                     {tt('userkeys_jsx.public_key.desc2')}{' '}
                                     <a
                                         className="public-keys__link"
-                                        href={`https://steemscan.com/@${account.get(
+                                        href={`https://steemscan.com/account/${account.get(
                                             'name'
                                         )}`}
                                         target="_blank"
                                         rel="noopener noreferrer"
                                     >
                                         Steemscan
+                                    </a>
+                                    <a
+                                        className="public-keys__link"
+                                        href={`https://steemdb.io/@${account.get(
+                                            'name'
+                                        )}`}
+                                        target="_blank"
+                                        rel="noopener noreferrer"
+                                    >
+                                        Steemdb
                                     </a>
                                 </p>
                             </div>
