+++ src/app/components/elements/WalletSubMenu.jsx
@@ -51,6 +51,22 @@ export default ({ accountname, isMyAccount, showTab }) => {
                     </Link>
                 </li>
             ) : null}
+            <li>
+                <Link
+                    to={`/@${accountname}/witnesses`}
+                    activeClassName="active"
+                >
+                    {tt('navigation.witnesses')}
+                </Link>
+            </li>
+            <li>
+                <Link
+                    to={`/@${accountname}/proposals`}
+                    activeClassName="active"
+                >
+                    {tt('g.proposals')}
+                </Link>
+            </li>
         </ul>
     );
 };
