+++ src/app/components/modules/ProposalList/styles.scss
@@ -1,30 +1,29 @@
 .ProposalsList {
-  max-width: 1000px;
   width: 100%;
   flex-direction: column;
   display: flex;
   padding-top: 20px;
 
   @media only screen and (min-width: 768px) {
-    width: 80%;
+    width: 100%;
   }
 
   @media only screen and (max-width: 768px) {
     width: 98%;
     padding: 0.5rem;
+    margin-left: auto;
+    margin-right: auto;
   }
-
-  margin-left: auto;
-  margin-right: auto;
 
   .search_proposal {
     margin: .75rem 0;
 
     input {
       border-color: transparent;
-
+      box-shadow: 0 0 5px rgba(109, 207, 246, 0.5);
       @include themify($themes) {
         background-color: themed('moduleBackgroundColor');
+        color: themed('textColorPrimary');
       }
     }
 
@@ -35,6 +34,10 @@
       @include themify($themes) {
         color: themed('colorAccent');
       }
+    }
+    input:focus {
+      border: 1px solid #06D6A9;
+      background: transparent;
     }
   }
 
@@ -133,7 +136,7 @@
   .proposals__item {
     @include themify($themes) {
       border-radius: themed('roundedCorners');
-      background-color: themed('moduleBackgroundColor');
+      background-color: themed('backgroundColorSecondary');
     }
 
     box-shadow: 1px 1px 5px 1px rgba(0, 0, 0, 0.05);
@@ -353,9 +356,15 @@
         transform: translateY(-50%);
       }
 
-      span {
+      span.vote-btn {
         cursor: pointer;
         margin-bottom: 2px;
+      }
+
+      span.vote-btn.disabled {
+        opacity: 0.8;
+        box-shadow: 0px 0px 0px 0 rgba(0, 0, 0, 0);
+
       }
 
       .total_votes {
