+++ src/app/components/pages/UserProfile.jsx
@@ -14,6 +14,8 @@ import UserKeys from 'app/components/elements/UserKeys';
 import PasswordReset from 'app/components/elements/PasswordReset';
 import CreateCommunity from 'app/components/elements/CreateCommunity';
 import Delegations from 'app/components/modules/Delegations';
+import Proposals from 'app/components/modules/Proposals';
+import Witnesses from 'app/components/modules/Witnesses';
 import UserWallet from 'app/components/modules/UserWallet';
 import Settings from 'app/components/modules/Settings';
 import CurationRewards from 'app/components/modules/CurationRewards';
@@ -70,6 +72,10 @@ export default class UserProfile extends React.Component {
     }
 
     componentDidUpdate(prevProps) {
+        const { accountname, getWithdrawRoutes } = this.props;
+        if (accountname && accountname !== prevProps.accountname) {
+            getWithdrawRoutes(accountname);
+        }
         this.redirect();
     }
 
@@ -150,7 +156,39 @@ export default class UserProfile extends React.Component {
                     />
                 </div>
             );
-        } else if (section === 'delegations') {
+        } else if (section === 'proposals') {
+            walletClass = 'active';
+            tab_content = (
+                <div>
+                    <div className="row">
+                        <div className="column">
+                            <WalletSubMenu
+                                accountname={account.name}
+                                isMyAccount={isMyAccount}
+                            />
+                        </div>
+                    </div>
+                    <br />
+                    <Proposals walletSectionAccount={accountname} />
+                </div>
+            );
+        }  else if (section === 'witnesses') {
+            walletClass = 'active';
+            tab_content = (
+                <div>
+                    <div className="row">
+                        <div className="column">
+                            <WalletSubMenu
+                                accountname={account.name}
+                                isMyAccount={isMyAccount}
+                            />
+                        </div>
+                    </div>
+                    <br />
+                    <Witnesses walletSectionAccount={accountname} />
+                </div>
+            );
+        }  else if (section === 'delegations') {
             walletClass = 'active';
             tab_content = (
                 <div>
